Change noise interface params

+++ src/main/java/net/minestom/server/utils/noise/Noise2D.java
@@ -2,5 +2,5 @@ package net.minestom.server.utils.noise;
 
 @FunctionalInterface
 public interface Noise2D {
-    double getValue(int x, int y);
+    double getValue(double x, double y);
 }
