+++ src/main/java/net/minestom/server/utils/noise/Noise3D.java
@@ -2,5 +2,5 @@ package net.minestom.server.utils.noise;
 
 @FunctionalInterface
 public interface Noise3D {
-    double getValue(int x, int y, int z);
+    double getValue(double x, double y, double z);
 }
